+++ lib/predicate/nodes/eq.rb
@@ -43,5 +43,9 @@ class Predicate
       900
     end
 
+    def evaluate(tuple)
+      left.evaluate(tuple) == right.evaluate(tuple)
+    end
+
   end
 end