+++ lib/predicate/nodes/gte.rb
@@ -6,5 +6,9 @@ class Predicate
       :>=
     end
 
+    def evaluate(tuple)
+      left.evaluate(tuple) >= right.evaluate(tuple)
+    end
+
   end
 end