+++ lib/predicate/nodes/identifier.rb
@@ -14,5 +14,9 @@ class Predicate
       @free_variables ||= [ name ]
     end
 
+    def evaluate(tuple)
+      tuple[name]
+    end
+
   end
 end