+++ lib/predicate/nodes/or.rb
@@ -6,5 +6,9 @@ class Predicate
       :'||'
     end
 
+    def evaluate(tuple)
+      sexpr_body.any?{|op| op.evaluate(tuple) }
+    end
+
   end
 end